add test guideline
