+++ examples/generation/run_deepseek_v2_lite_math.sh
@@ -1,7 +1,7 @@
 set -x
 
-data_path=$HOME/data/rlhf/gsm8k/test.parquet
-save_path=$HOME/data/rlhf/math/deepseek_v2_lite_gen_test.parquet
+data_path=$HOME/data/gsm8k/test.parquet
+save_path=$HOME/data/gsm8k/deepseek_v2_lite_gen_test.parquet
 model_path=deepseek-ai/deepseek-llm-7b-chat
 
 python3 -m verl.trainer.main_generation \
