+++ tests/distributed/README.md
@@ -1,0 +1,1 @@
+This folder is reserved for unit tests (instead of end-to-end tests) that require multiple GPUs.
