+++ tests/e2e/README.md
@@ -1,0 +1,1 @@
+This folder is reserved for end-to-end tests that typically require multiple GPUs.
