+++ tests/standalone/README.md
@@ -1,0 +1,1 @@
+The standalone test folder is reserved for tests that require dedicated environment (e.g. memory stress tests)
