+++ tests/utils/test_linear_cross_entropy.py
